refactor standardizer
